test case for socks udp
